pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dbt_affiliate_api"
version = "0.1.0"
description = "Affiliate API ingestion project with dbt-duckdb extraction, header staging tables, and parquet export."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "dbt-duckdb>=1.8.0",
  "duckdb>=1.1.0",
  "Jinja2>=3.1.0",
  "pandas>=3.0.1",
  "pyarrow>=23.0.1",
  "PyYAML>=6.0.0",
  "requests==2.32.5",
]

[project.scripts]
affiliate-api = "python_utils.cli:main"

[tool.setuptools.packages.find]
where = ["."]
